requirements.txt
Django==2.2.28
nats-py==2.10.0
docopt==0.6.2